[notice]: this feature branch now compiles successfully. The coupled temperature at the interface is only avaialable when the following fluid and solid models are selected:
1- fluidModel: buoyantBoussinesqPimpleFluid
2- solidModel: thermalLinGeomSolid

This has not been tested with any tutorial involving coupled temperature yet,
so I don't know if the solver actually runs. The good news is, it doesn't break
anything and default tutorials are unaffected.

It is straightforward to expand the capabality to other models, but for now
I'm waiting for the critical bug in the fluidModels to be fixed by Dr.Cardiff
and I can actully test a new tutorial with coupled temperature field.